forgot to populate author and category id to getPost function

+++ controllers/posts/posts.js
@@ -95,8 +95,11 @@ exports.getAllPosts = asyncHandler(async (req, res) => {
 //@access PUBLIC
 
 exports.getPublicPosts = asyncHandler(async (req, res) => {
-  const posts = await Post.find({}).sort({ createdAt: -1 }).limit(4);
-  res.status(200).json({
+  const posts = await Post.find({})
+    .sort({ createdAt: -1 })
+    .limit(4)
+    .populate("category");
+    res.status(200).json({
     status: "success",
     message: "Posts successfully fetched",
     posts,
@@ -108,7 +111,9 @@ exports.getPublicPosts = asyncHandler(async (req, res) => {
 //@access Public
 
 exports.getPost = asyncHandler(async (req, res) => {
-  const post = await Post.findById(req.params.id);
+  const post = await Post.findById(req.params.id)
+  .populate("author")
+  .populate("category");
   res.status(201).json({
     status: "success",
     message: "Post successfully fetched",
@@ -137,6 +142,12 @@ exports.updatePost = asyncHandler(async (req, res) => {
 //@access Private
 
 exports.deletePost = asyncHandler(async (req, res) => {
+  // find the post
+  const postFound = await Post.findById(req.params.id);
+  const isAuthor = req.userAuth?._id.toString() === postFound?.author?._id.toString();
+  if (!isAuthor) {
+    throw new Error("Action denied, You are not the author of this post");
+  }
   await Post.findByIdAndDelete(req.params.id);
   res.status(201).json({
     status: "success",
